requirements.txt
matplotlib==3.0.2
Keras==2.3.1
numpy==1.16.0
Pillow==6.2.1
